Fix responsive for editing tempaltes inside post / page
